requirements.txt
langchain
langchain-community
langchain-groq
streamlit
python-dotenv
pypdf
pandas
openpyxl
sentence-transformers<3.0
langchain-huggingface
langchain-text-splitters
tiktoken
supabase
Pillow
streamlit-paste-button
